chore(ARCH-482/Notification): remove deprecated lifecycle and use RTL for tests
