+++ packages/components/src/Notification/Notification.test.js
@@ -1,200 +1,223 @@
 import React from 'react';
-import { mount } from 'enzyme';
+import { render, screen, fireEvent } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+
 import Notification from './Notification.component';
 
-jest.useFakeTimers();
-
-describe('Message', () => {
-	it('should render an article with one paragraph if notification is not an array', () => {
-		const notification = { message: 'bar' };
-		const wrapper = mount(<Notification notifications={[notification]} leaveFn={jest.fn()} />);
-		expect(wrapper.find('article').length).toBe(1);
-		expect(wrapper.find('p').length).toBe(1);
-	});
-	it('should render an article if notification is an array', () => {
-		const notification = { message: ['foo', 'bar'] };
-		const wrapper = mount(<Notification notifications={[notification]} leaveFn={jest.fn()} />);
-		expect(wrapper.find('article').length).toBe(1);
-		expect(wrapper.find('p').length).toBe(2);
-	});
-});
-
-describe('TimerBar', () => {
-	it('should render not time bar if type === error', () => {
-		const notification = { message: 'bar', type: 'error' };
-		const wrapper = mount(<Notification notifications={[notification]} leaveFn={jest.fn()} />);
-		expect(wrapper.find('div.tc-notification-timer-bar').length).toBe(0);
-	});
-	it('should render a TimeBar', () => {
-		const notification = { message: 'bar', type: 'success' };
-		const wrapper = mount(<Notification notifications={[notification]} leaveFn={jest.fn()} />);
-		expect(wrapper.find('div.tc-notification-timer-bar').length).toBe(1);
-	});
-});
-
 describe('Notification', () => {
-	it('should render error', () => {
-		const notification = {
-			message: 'foo',
-			type: 'error',
-			onClick: jest.fn(),
-			onMouseEnter: jest.fn(),
-			onMouseOut: jest.fn(),
-		};
-		const leaveFn = jest.fn();
-		const wrapper = mount(<Notification notifications={[notification]} leaveFn={leaveFn} />);
-		expect(wrapper.find('div.tc-notification').getElement()).toMatchSnapshot();
-	});
-});
-
-describe('NotificationContainer', () => {
-	describe('Timer Registry Utility', () => {
-		it('register, pause, resume, cancel, isRegistered', () => {
-			const notification = {
-				id: 'test-1',
-			};
-			const mockTimer = {
-				pause: jest.fn(),
-				resume: jest.fn(),
-				cancel: jest.fn(),
-			};
-			const instance = new Notification({ notifications: [] });
-			instance.registry.register(notification, mockTimer);
-			expect(instance.registry.isRegistered(notification)).toEqual(true);
-			instance.registry.pause(notification);
-			expect(mockTimer.pause).toHaveBeenCalled();
-			instance.registry.resume(notification);
-			expect(mockTimer.resume).toHaveBeenCalled();
-			expect(instance.registry.timerRegistry).toHaveProperty(notification.id);
-			instance.registry.cancel(notification);
-			expect(mockTimer.cancel).toHaveBeenCalled();
-			expect(instance.registry.timerRegistry).toEqual({});
-		});
+	afterEach(() => {
+		jest.useRealTimers();
 	});
 
-	describe('Event Handler', () => {
-		let mockEvent;
-		let mockStopPropagation;
-		let mockGetAttribute;
-		let mockSetAttribute;
-		let mockLeaveFn;
-		const mockNotification = {
-			type: 'info',
-		};
-
-		beforeEach(() => {
-			mockStopPropagation = jest.fn();
-			mockGetAttribute = jest.fn(() => '');
-			mockSetAttribute = jest.fn();
-			mockLeaveFn = jest.fn();
-			mockEvent = {
-				currentTarget: {
-					setAttribute: mockSetAttribute,
-					getAttribute: mockGetAttribute,
+	it('should render', () => {
+		// given
+		const notifications = [
+			{
+				id: 'story-1',
+				message: 'This is a feedback of your operation1, This is a feedback of your operation1',
+				action: {
+					label: 'Haha',
+					icon: 'talend-undo',
+					onClick: () => {},
 				},
-				stopPropagation: mockStopPropagation,
-			};
-		});
-
-		it('onMouseEnter', () => {
-			const instance = new Notification({ notifications: [] });
-			instance.registry.pause = jest.fn();
-			const event = {
-				currentTarget: {
-					setAttribute: jest.fn(),
-				},
-			};
-			instance.onMouseEnter(event, mockNotification);
-			expect(instance.registry.pause).toHaveBeenCalledWith(mockNotification);
-		});
-
-		it('onMouseOut', () => {
-			const props = {
-				leaveFn: jest.fn(),
-				autoLeaveTimeout: 4000,
-				notifications: [],
-			};
-			const instance = new Notification(props);
-			instance.registry.register = jest.fn();
-			instance.registry.resume = jest.fn();
-			instance.onMouseOut(mockEvent, mockNotification);
-			expect(mockEvent.currentTarget.getAttribute).toHaveBeenCalledWith('pin');
-			expect(instance.registry.resume).toHaveBeenCalledWith(mockNotification);
-		});
-
-		it('onClick', () => {
-			const props = { notifications: [], leaveFn: jest.fn() };
-			const instance = new Notification(props);
-			instance.onClick(mockEvent, mockNotification);
-			expect(mockGetAttribute).toHaveBeenCalledWith('pin');
-			expect(mockSetAttribute).toHaveBeenCalledWith('pin', 'true');
-			mockEvent.currentTarget.getAttribute = jest.fn(() => 'true');
-			instance.onClick(mockEvent, mockNotification);
-			expect(mockSetAttribute.mock.calls[1]).toEqual(['pin', 'false']);
-			expect(props.leaveFn).toHaveBeenCalledWith(mockNotification);
-		});
-
-		it('onClose', () => {
-			const props = {
-				leaveFn: mockLeaveFn,
-				notifications: [],
-			};
-			const instance = new Notification(props);
-			instance.registry.cancel = jest.fn();
-			instance.onClose(mockEvent, mockNotification);
-			expect(mockStopPropagation).toHaveBeenCalled();
-			expect(instance.registry.cancel).toHaveBeenCalledWith(mockNotification);
-			expect(mockLeaveFn).toHaveBeenCalled();
-		});
-	});
-
-	describe('Autoleave', () => {
-		let notifications;
-		const mockLeaveFn = jest.fn();
-
-		beforeEach(() => {
-			notifications = [
-				{
-					id: 0,
-					message: 'This is a feedback of your operation1, This is a feedback of your operation1',
-					action: {
-						label: 'UNDO',
-						icon: 'talend-undo',
-						onClick: () => {},
-					},
-				},
-				{
-					id: 1,
-					type: 'info',
-					message: 'This is a feedback of your operation1, This is a feedback of your operation1',
-				},
-				{
-					id: 2,
-					type: 'warning',
-					message: 'This is a feedback of your operation2, This is a feedback of your operation2',
-				},
-				{
-					id: 3,
-					type: 'error',
-					message: 'This is a feedback of your operation3, This is a feedback of your operation3',
-				},
-			];
-		});
-
-		it(
-			'should register autoLeave only for notification of type info or warning ' +
-				'should trigger leaveFn after timeout',
-			() => {
-				const props = {
-					notifications,
-					leaveFn: mockLeaveFn,
-				};
-				const instance = new Notification(props);
-				expect(instance).toBeDefined();
-				expect(mockLeaveFn).not.toHaveBeenCalled();
-				jest.runAllTimers();
-				expect(mockLeaveFn).toHaveBeenCalledTimes(3);
 			},
+			{
+				id: 'story-2',
+				type: 'error',
+				message: [
+					'This is a feedback of your operation2',
+					'This is a feedback of your operationZ, This is a feedback of your operationZ',
+				],
+				action: {
+					label: 'undo',
+					icon: 'talend-undo',
+					onClick: () => {},
+				},
+			},
+			{
+				id: 'story-3',
+				type: 'warning',
+				message: ['This is a feedback of your operation3', 'details'],
+			},
+			{
+				id: 'story-4',
+				type: 'warning',
+				message: 'This is a feedback of your operation4',
+			},
+		];
+
+		// when
+		render(<Notification notifications={notifications} leaveFn={() => {}} />);
+
+		// then
+		expect(screen.getAllByRole('status').length).toBe(3);
+		expect(screen.getAllByRole('alert').length).toBe(1);
+		notifications
+			.map(({ message }) => message)
+			.flat()
+			.forEach(message => {
+				expect(screen.getByText(message)).toBeInTheDocument();
+			});
+	});
+
+	it('should render an Action in notification', () => {
+		// given
+		const action = { id: 'lol', onClick: jest.fn(), 'data-testid': 'lol' };
+		const notification = { id: 'id', message: 'foo', action };
+
+		// when
+		render(<Notification notifications={[notification]} leaveFn={jest.fn()} />);
+
+		// then
+		expect(screen.getByTestId('lol')).toBeInTheDocument();
+	});
+
+	it('should render a simple message', () => {
+		// given
+		const notification = { id: 'id', message: 'bar' };
+
+		// when
+		render(<Notification notifications={[notification]} leaveFn={jest.fn()} />);
+
+		// then
+		expect(screen.getByText('bar')).toBeInTheDocument();
+	});
+
+	it('should render an an array of message', () => {
+		// given
+		const notification = { id: 'id', message: ['foo', 'bar'] };
+
+		// when
+		render(<Notification notifications={[notification]} leaveFn={jest.fn()} />);
+
+		// then
+		expect(screen.getByText('foo')).toBeInTheDocument();
+		expect(screen.getByText('bar')).toBeInTheDocument();
+	});
+
+	it('should render a time bar', () => {
+		// given
+		const notification = { id: 'id', message: 'foo' };
+
+		// when
+		render(<Notification notifications={[notification]} leaveFn={jest.fn()} />);
+
+		// then
+		expect(screen.getByTestId('timer')).toBeInTheDocument();
+	});
+
+	it('should render not time bar with error notification', () => {
+		// given
+		const notification = { id: 'id', message: 'foo', type: 'error' };
+
+		// when
+		render(<Notification notifications={[notification]} leaveFn={jest.fn()} />);
+
+		// then
+		expect(screen.queryByTestId('timer')).not.toBeInTheDocument();
+	});
+
+	it('should pin notification when the user click', () => {
+		// given
+		const leaveFn = jest.fn();
+		const notification = { id: 'id', message: 'foo' };
+		render(<Notification notifications={[notification]} leaveFn={leaveFn} />);
+		const notifDiv = screen.getByRole('status');
+		expect(notifDiv).not.toHaveAttribute('pin', 'true');
+
+		// when
+		userEvent.click(notifDiv);
+
+		// then
+		expect(notifDiv).toHaveAttribute('pin', 'true');
+	});
+
+	it('should call leaveFn props when the user click', () => {
+		// given
+		const leaveFn = jest.fn();
+		const notification = { id: 'id', message: 'foo' };
+		render(<Notification notifications={[notification]} leaveFn={leaveFn} />);
+		expect(leaveFn).not.toHaveBeenCalled();
+		const notifDiv = screen.getByRole('status');
+
+		// when
+		userEvent.click(notifDiv); // pin
+		userEvent.click(notifDiv);
+
+		// then
+		expect(leaveFn).toHaveBeenCalledWith(notification);
+	});
+
+	it('should call leaveFn props when the user click on close button', () => {
+		// given
+		const leaveFn = jest.fn();
+		const notification = { id: 'id', message: 'foo' };
+		render(<Notification notifications={[notification]} leaveFn={leaveFn} />);
+		expect(leaveFn).not.toHaveBeenCalled();
+		const closeBtn = screen.getByRole('button');
+
+		// when
+		userEvent.click(closeBtn);
+
+		// then
+		expect(leaveFn).toHaveBeenCalledWith(notification);
+	});
+
+	it('should call leaveFn after timeout', () => {
+		// given
+		jest.useFakeTimers();
+		const notification = { id: 'id', message: 'foo' };
+		const leaveFn = jest.fn();
+
+		// when
+		render(
+			<Notification notifications={[notification]} leaveFn={leaveFn} autoLeaveTimeout={2000} />,
 		);
+		expect(leaveFn).not.toBeCalled();
+		jest.advanceTimersByTime(2000);
+
+		// then
+		expect(leaveFn).toBeCalled();
+	});
+
+	it('should not call leaveFn after timeout on error', () => {
+		// given
+		jest.useFakeTimers();
+		const notification = { id: 'id', message: 'foo', type: 'error' };
+		const leaveFn = jest.fn();
+
+		// when
+		render(
+			<Notification notifications={[notification]} leaveFn={leaveFn} autoLeaveTimeout={2000} />,
+		);
+		expect(leaveFn).not.toBeCalled();
+		jest.advanceTimersByTime(2000);
+
+		// then
+		expect(leaveFn).not.toBeCalled();
+	});
+
+	it('should pause/resume timer on hover/leave', () => {
+		// given
+		jest.useFakeTimers();
+		const notification = { id: 'id', message: 'foo' };
+		const leaveFn = jest.fn();
+		render(
+			<Notification notifications={[notification]} leaveFn={leaveFn} autoLeaveTimeout={2000} />,
+		);
+
+		// when
+		fireEvent.mouseEnter(screen.getByText('foo'));
+		jest.advanceTimersByTime(2000);
+
+		// then
+		expect(leaveFn).not.toBeCalled();
+
+		// when
+		fireEvent.mouseOut(screen.getByText('foo'));
+		jest.advanceTimersByTime(2000);
+
+		// then
+		expect(leaveFn).toBeCalled();
 	});
 });
